fix: 26505 Losing focus on the amount input field when switching between apps
